merge `scr` and `scr2` scrolling registers
`scr` affected the children of a given widget, while `scr2` affected built-in components of widgets, like menu items.
I believe the separation started during initial work on menu widgets. I kept it in because it seemed useful for anchoring
child widgets within a parent that also had scrolling built-in components. For example, a scrolling menu widget of save
files could feature a "Load" child button anchored near the edge of its body. As of this writing, no such usage has ever
occurred in any ProdUI widgets. The "load" button could be hard-coded as part of the menu widget, or a container widget
could hold both the menu and a separate "load" button as its children.

Two sets of scroll registers give the impression that they can both work independently in the same widget. This was
not the case: with the current widget scrolling API, only one or the other could be active in a single widget at a
time. Supporting both sets required numerous support functions and methods to be doubled for very little gain.

Now, `scr` affects both child widget offsets and the positioning of built-in components.