Rename callback 'uiCall_pointerHoverMove' to 'uiCall_pointerHover'
This callback originally emitted only when the mouse cursor moved. It
was changed temporarily to fire on every love.update() as well, to
resolve the problem of widgets moving out from underneath an idle mouse
cursor.

This "temporary" behavior has been implemented for so long that the
callback should just be renamed.