Mouse Wheel scrolling: swallow events when scroll registers are not settled
This commit is hard to explain in one line.

The user is interacting with a container that is taller than the screen.
When flicking the mouse wheel (while the cursor is not over any child
widget), the container view scrolls vertically.

In this container are ListBoxes (or other similar menu widgets). When
flicking the wheel over these widgets, they take priority over their
parent container, and try to scroll their own content while preventing
the container from scrolling. If, however, the child menu is already
scrolled as far as it can go, then it yields the event to the parent.

The problem is that the scroll-to animation can still be playing while
the target scroll position is already at the minimum or maximum
position. To the user, the container appears to scroll prematurely.

This commit changes the yield conditions so that the child does not
cede the event when its target scroll position and its actual scroll
position differ.

In practice, this means that if you flick the wheel over a menu while
being close to start/end, some of those love.wheelmoved events will
not make it to the container widget.

This behavior depends on the animation speed, and it may need tweaking
or replacement in the future.