Separate widget creation and initialization steps
`context:addWidget()` and `widget:addChild()` no longer accept a
table with pre-written state to be used as the basis for the new
widget. They now return the widget in an uninitialized state, and
the library user must complete the setup by calling
`widget:initialize()`.

Reasons for this change:

1) A pre-written table passed as an argument leaves open the
   possibility of someone sharing tables, leading to duplicate
   tables as widget instances.

2) I need more control over the widget setup process. Specifically,
   this lets me assign layout state before full initialization.

The callback `uiCall_create` is renamed to `uiCall_initialize`. It
now accepts a varargs list, and it no longer bubbles up to the root.

The user event `userCreate` is now `userInitialize`.

The base widget metatable contains default values for `x`, `y`, `w`
and `h`.