Merge menu sub-objects into widgets
Menu state that was contained in `self.menu` has been moved up
one level (to the main widget table), and menu methods are now
attached to the widget def.

Menus were implemented as a standalone structure so that one could
attach the same menu to multiple widgets. In practice, this was only
used for the drawers that were invoked by Dropdown and ComboBox
widgets, and even in those cases, it might be safer to just recreate
the menu contents.