More work on merging reshape and layout procedures
This batch of commits has broken some stuff; the reshape code
for most control widgets needs review.

Basically, reshaping and layout have been merged into one action.
I split the 'uiCall_reshape' event into multiple callbacks. One
set of callbacks is fired before and after the layout is applied,
and another set is fired before and after the widget's children
are handled. This is more complicated, but it should allow special
widgets, which vary in size along one axis depending on their
content, to make some necessary dimensional changes before the
layout is applied. An example of such a widget is a block of
text that grows and shrinks vertically when its content is
changed, or when its parent container is narrowed or expanded.

The WIMP Demo categories are now listed in a TreeBox, and I need
to make some panels for the branch nodes (which will be broad
categories, like "controls").