WIMP Demo: add zoom for troubleshooting
The feature can be turned on by setting 'demo_zoom_enable'
to true. (I haven't set up a control to toggle it yet.) Then
press '-' and '=' to zoom in and out. The centering and rate
of accumulation/decumulation are both out of whack, but it
works well enough to inspect borders and margins.