Clicking containers sends thimble1 to UI Frame
Translation: When you click on the empty space in a container, it
now searches for the UI Frame (Window or Workspace) that owns it,
and attempts to transfer over the thimble (keyboard focus).
Previously, it was trying to take the thimble for itself, even
though it's no longer programmed to be capable of having it.

In earlier commits, 'base/container' was *the* container widget for
other controls. Window Frames had an embedded container, and
Workspaces didn't exist yet. The structure has changed, and now
those second-gen UI Frames are where the thimble should be parked
when the user clicks on "nothing."