Deprecate 'WIP.md'
This file contained a roadmap of pending tasks. The objective was to
develop all WIMP widgets, without regard for other concerns (unless
they directly impeded widget work). This does not reflect my
objective of the past month or so, which is to improve the library's
internal structure and its overall presentation.

If this file makes a comeback, it will likely be in of the docs repo,
or be integrated into a blog or forum post.