WIP: replace sash with divider
I'd like to make the main demo ListBox resizable, but the existing
sash widget is just not cutting it. I am replacing it with a new
'divider' container, where widgets are assigned to nodes in a tree,
and the container figures out the sizing and positioning.

I have some basic layout code working, but the actual resize action
is not yet implemented (it needs to be cannibalized from the old
sash widget).