Divider and sash work
In the divider test, the separator between boxes (F) and (G) is
a prototype sash. This is not a widget itself, but rather a special
layout node that is managed by the divider. The sash intersection
rectangle is condensed before the mouse cursor enters, and expanded
before it leaves. For testing purposes, these settings are cranked
up to the point of being annoying.

Sashes are limited to operating on slice nodes measured in pixels
(and not percentages).