More theme work
I made a mess of things, and it took a while to get the project
booting again. Some gross hacks remain in the theme handling code.
Scaling is still not properly implemented.

Long story short: ProdUI themes were originally envisioned as
having a lot of autonomy. The very first version of the Vacuum
theme contained all render code for all widgets, in addition to
colors and other rendering options. The functions were moved out
a while ago, but before these recent commits, the theme was still
mutating numbers (clamping, scaling, flooring) as part of chunk
execution. I want themes to be mostly data, so that they can be
tweaked, modified, and layered with other data before being
applied to the context.