Theme work
Rather than reinvent the wheel, I might as well take advantage of
PhysicsFS's ability to mount multiple archives to one path. (You
can mount a ZIP file in the game directory by first loading it
as a FileData, and then passing that FileData reference to
love.filesystem.mount().)

After a few dead ends, I ended up with a function that loads a
directory of files as one Lua table. The UI theme is structured
this way; the idea is that you zip it, mount the zip(s), and then
amalgamate the path's contents into one table. Make adjustments to
the table as necessary, and then apply the theme to the UI context.
Heavier files, such as images and font data, are kept in a separate
path.

The handover of data from theme to context still isn't finalized.

Also, I moved the theme building system outside of the main prod_ui
directory, and added a subdirectory of temporary work files to the
project's .gitignore.