Further improve error messages in uiTheme
'uiTheme.error()' will prepend errors with some helpful
information, including the skinner ID (implementation), the
skin name (data), and a stack of keys that the checker functions
were in the middle of exploring when the error occurred.

The stack of keys is pushed and popped manually within the checker
functions. This is a pain in the neck to set up, but it works
well enough.

The 'validate' and 'transform' functions are active in this commit.
They are still missing a few checks here and there. Further, I will
likely be disabling them to ease development; they shall be enabled
when the library is finished and properly released.