text box (single): fix caret position after call to setText()
Fixing this relatively benign problem required a large refactor
of the code for single-line text input. In short: some text box
methods did not properly update the state of the client widget
after they had finished executing. Functions bound to user key
input were fine, but anything called as part of user code, like
overwriting text as a result of the press of a UI button, was not
guaranteed to do the proper cleanup.