Fix unclickable scroll bars in Window Frames
This affected only scroll bars that were deployed as part of
Window Frames, and not those that are embedded within sub-controls
(ie the scroll bar belonging to the ListBox in the Widget Tree
viewer).

It was caused by mixing up absolute and relative coordinates in an
intersection test.

This regression has been present since at least April 11th. It
wasn't noticed because Window Frame scroll bars are infrequently
used in the WIMP demo (and when present, I would typically use the
mouse wheel to scroll, rather than interact with the bar directly.)