Fix scroll bar rendering at export DPIs greater than 96
Scroll bar troughs are drawn with a function that expects a
single white pixel, as defined by a 1x1 quad. The theme baking
system was scaling this quad according to the DPI export
setting. Since the quad was no longer 1x1, the trough would
be drawn larger than intended.

Fixed by adding a 'no_scale' table to the texture base_data,
which forces any such textures to be exported at 96 DPI (1.0).