Add Icon Sets to theme; update icon code in ListBox, TreeBox
The theme now provides sets of icons for a widget to choose from.

Fixed ListBoxes and TreeBoxes to update their items' icon
references when the theme is reloaded. A few more instances of
bad behavior need to be cleaned up (search 'resources.quads["atlas"]'
to see).

Also, 'common/common_tree.lua' was moved and renamed to
'shared/lgc_tree.lua'.