Theme selection at run time
* Rewrote the theme loader code to 1) accept either a single Lua
file or a directory of files (was previously just the latter),
and 2) support the patching of themes at the time of loading them.
This simplifies making variations on themes. For example, the
light version of Vacuum Dark adjusts colors and a texture
reference, but it otherwise doesn't alter things like sizes.

* Added rxi's json.lua to the project. I will probably be changing
the theme tables to be stored on disk from Lua source to a non-
interactive format; JSON seems to be the most likely candidate.

* Finished up the first pass of theme colors and SVGs for Vacuum
Light. Several issues remain, such as white text that needs to
be made darker.