Multi-Line text input work
The multi-line input code hasn't been touched in a long time.
I need to bring it up to date with the equivalent single-line
programming. It's much more complicated, though, so I think
it's going to take me a while.

This commit fixes one error with 'Select All' (at best, it didn't
select to the far end; at worst, it crashed).