Text input: fix shift+click to select range
When you click on editable text while holding the shift key, the
expected behavior is to highlight from the previous caret position
to the character that is now under the mouse cursor. The logic to
check the state of the shift key was missing.

This affected both the single-line and multi-line text box widgets.