Merge Widget:initialize() with Widget:addChild() / Context:addRoot()
The 'Widget:initialize()' method allowed the developer to configure
widget properties before executing its 'uiCall_initialize()' callback.
It replaced an earlier method of passing a pre-configured table, which
would then be used as the basis for the new widget instance. Both
approaches were used mainly to specify skin IDs, and to furnish
necessary data like LÖVE Fonts.

This commit eliminates the explicit call to 'Widget:initialize()',
and moves the invocation of 'uiCall_initialize()' to
'Widget:addChild()' and 'Context:addRoot()'. To set widget data ahead
of time, these two methods accept an argument for the skin ID, and a
varargs list which is passed to 'uiCall_initialize()'.

The previous two ways of preconfiguring widgets introduced the
possibility of accidental double initializations. What small benefit
the "explicit init" provided was overshadowed by the confusion and
annoyance of forgetting to call the stupid method for every single
widget instance. Finally, if a widget really needs an init step, it
can just be made part of the widget definition.