(bugged) Multi-line text refactor
Restructured the multi-line code to (more or less) be like its
single-line counterpart.

I still need to fix and review. Interacting with the multi-line
text box in this commit will lead to a crash.

Also, I removed glyph masking from the multi-line code because it
couldn't mask line feeds. That's more of a single-line thing,
anyway; a better solution for multi-line might be to just not
render text at all when a 'hide_contents' flag is set.