Change location + implementation of keybindings for text input
This will probably be changed further. The important part is that
the keybindings are no longer stored in the input library code,
but rather with other ProdUI user settings.

I have changed the key lookup from a simple hash of keys to a dumb,
linear search through an array of arrays. This is temporary, I
swear! The problem with the hash format is that it prioritizes
hotkeys over the actual command IDs, which is bad from a user
perspective.

 [C+y] = "redo",
 [CS+z] = "redo"

VS

 {"redo", "C+y", "CS+z"}

We can have our cake and eat it too with only a small amount of
overhead, by generating a hash from the array. But that needs to
be synchronized with every change, and I have not yet decided where
such things should be stored.