NumberBox: fix errors with undo/redo history
Also, general cleanup of this widget and some supporting code.

I modified NumberBox's 'fn_check' method, which is supposed to
accept or reject new text box input, to also modify the line
under certain circumstances. This feels wrong, but I'm not sure
where else to put this code so that it kicks in for every form
of input (key-press, text input, commands issued from pop-up
menus).