WIP: restructure project
I added an 'init.lua' to the project, which returns a table with
all of the main source files that begin with 'ui_'. The library
user could require (load) the ProdUI directory as the local
variable 'ui', and from there, they could access 'uiAssert' as
'ui.assert', 'uiGraphics' as 'ui.graphics', and so on. I am not
100% sure that I'll go ahead with this.

In any case, it would be good to make some more 'ui_' files so
that library users don't have to dig deep through the ProdUI source
to get what they need (or worse, it's there and they never learn
about it).