Merge layout nodes and widgets
(Note: sashes are broken in this commit.)

I've changed how widget layouts are specified. Parent widgets
used to have a separate tree of nodes, which offered some
nice flexibility. Sashes were also implemented as special
nodes.

Unfortunately, it was a colossal pain in the neck to work
with. Now, instead of a tree of nodes, the parent just
maintains a separate array of children which are part of its
layout. Hierarchical designs, certain kinds of gaps, sashes,
etc., now require widgets.