Get sashes working again; change behavior of dead widgets
The sashes need to be brought up to a "professional"
standard, but they are now at least functional.

I changed the widget removal method to strip the
metatable. This exposed some problems in ProdUI's mouse
event code, which I have patched up with additional
checks for the field 'wid._dead'.