WIP: Rewrite sashes again
Sashes are broken (and disabled) in this commit.

I overlooked the problem of sashes being dragged too
far to the right or the bottom, becoming impossible
to click. I plan to change them from widgets to
sensors in the parent container—*again*—with some
additional code to carve out space for sashes in
the layout space.

I also neglected to think about centering sashes
(ie cutting a bit of the widget segment to make room
for half of the sash).