Add generic Rectangle structure
ProdUI has a few axis-aligned rectangular structures that are
implemented as arbitrary fields in an owner object. This commit
adds a new module, 'pile_rectangle.lua', for manipulating
rectangle tables, and it changes widget viewports to work with
that module.

Before:

-- Viewport #1: self.vp_x, self.vp_y, self.vp_w, self.vp_h
-- Viewport #2: self.vp2_x, self.vp2_y, self.vp2_w, self.vp2_h
-- …and so on, up to 8.

After:

-- local vp, vp2 = self.vp, self.vp2
-- Viewport #1: vp.x, vp.y, vp.w, vp.h
-- Viewport #2: vp2.x, vp2.y, vp2.w, vp2.h
-- …and so on, up to 8.

The ultimate goal is to use the same set of rectangle
manipulation functions for a variety of rectangular internals.
'pile_rectangle.lua' is mostly based on functions that worked
only on widget viewports (and they contained a bunch of
indirection to get the correct keys for a given viewport; that
stuff is thankfully excised).

The reason for viewports previously being arbitrary fields is a
(probably overblown) concern for allocating too many tables per
widget. I have set up a resource pool for giving and taking
Rectangles, but at this point, I don't know if it is actually
faster than just letting the garbage collector do its thing.