Upgrade PILE Base to 1.300; change Enum declarations
This is prep work for adding and testing a table validator,
to be used for theme and skin data. The current means of
checking skin data is not able to account for extra,
superfluous keys.

I also hoped to derive skin data properties from the
validation tables, but I'm not sure how practical that
will be.