Schema implementation work
This commit changes how theme resources are validated.
The previous implementation was based around function calls
and a manually managed stack of labels. It was probably a
tiny bit faster than the new data-oriented approach, but it
didn't catch unused keys in skin data. It also used a
complicated set of assertion functions, which relied on a
special error handler to concatenate the label stack and
error message.

The new system is built around 'pcall'ing assertions that
follow the PILE Argument Check function signature. Most of
my effort on this commit has been focused on rewriting PILE
Schema to find the right balance of features, and to make a
notation that isn't wretched to type out by hand.

This commit still has some bugs (for example, I haven't
dealt with the light theme yet), so I will not be pushing to
remote until those are cleared up. I also need to make a
formal update to PILE Base.