wimp/button_split: fix mouse-click collision detection
It was possible to click the edges of the "auxiliary function"
area and trigger the primary button action.

Cause: the code was incorrectly written to use the "center this
graphic" viewport, which had margins carved out, for the purpose
of detecting mouse clicks in the auxiliary zone.

Fixed by adding another viewport to handle click checks.