WIMP Demo, debug panel: eliminate flicker
When summoning the debug information panel with
Ctrl+Shift+1 or Ctrl+Shift+2, if the mouse cursor was
directly overhead, the panel would visibly and irritatingly
jump to the other side of the window after one frame.

Moved some things around so that the position change occurs
before the panel is rendered.