wimp/menu_tab: add alignment for header text in columns
Center alignment looks bad, due to reserving space for
the sorting arrow. I tested not reserving it when an
arrow is *not* active in a column box, but it's
visually jarring when the arrow does appear (the header
text jumps back a bit).

I'm not sure how to proceed, so I'll leave it alone for
now.