Move generic data structure code to PILE Base
I am moving some code (related to managing data structures)
out of ProdUI and into PILE Base, a Lua utility library.

The structures are:

* Trees
* Doubly linked lists
* "HookLists" (callable arrays of functions)

This change is taking a while to complete because I've
been fretting about the details.

As of this commit, the new beta version of PILE Base is
integrated, and the WIMP demo is booting. I need to search
the demo for bugs, then push a formal update to PILE Base,
then update the version of PILE Base in ProdUI, and so on.

Afterwards, I'd like to rename some fields and methods.