Improve checks for dead widgets in thimble code
Opening a ComboBox drawer and then clicking away from it
would crash the WIMP demo.

Maybe it would be better to forbid destroying widgets
within those code paths?
In the previous commit, it was possible to crash the WIMP
demo by opening the ComboBox drawer and then clicking
outside of it.