input/script_editor: add line number column
I had this implemented in LineEd ages ago, back in the 'DemoUI'
days. I dropped it at the time because I felt like it was
getting in the way.

There remains a problem where the vertical scrolling "hitches"
when the user is creating new lines at the end of the document,
and an additional digit happens to be added to the line number
column. It has to do with the order of how things are updated.
The bug can be mitigated somewhat by setting the reserved
digits count to 4 (such that it will only begin to happen upon
reaching the ten thousandth line).